F53OSC Encryption (#39)
Implement encryption with F53OSC. Note that this is F53OSC-specific, not based on the OSC spec. The core of the encryption is written in Swift using CryptoKit, including P521 public/private keys, HKDF derived symmetric keys, and ChaChaPoly symmetric encryption.

In addition to the cryptographic code, this includes a new encryption handshake protocol, which can be transmitted over a new F53OSC control message channel (which looks a lot like OSC to avoid duplicating code, but starts with `!` instead of `/` to differentiate it from OSC messages. Also, a new encrypted message type is added, which consists of encrypted OSC message data prefixed by a `#` character.

Clients wishing to use encryption support encryption must first generate a key pair (using the `generateKeyPair` method of `F53OSCEncrypt`) and then call `connectEncryptedWithKeyPair:` on their `F53OSCClient`.

Servers wishing to support encryption must first generate a key pair (using the `generateKeyPair` method of `F53OSCEncrypt`) and then call `setKeyPair:` on their `F53OSCServer`.

Additionally, both servers and clients can check the `encrypter.publicKey` property of an `F53OSCSocket` to verify the identity of a peer if they wish to do so.